// ISS/web-final/src/enums/Route.ts
export enum Route {
  Login = "/login",
  Home = "/home",
  Grades = "/grades",
  Curriculum = "/curriculum",
  Courses = "/teacher/courses",
  ChiefCourses = "/departmentChief/courses",
  OptionalCourses = "/optional-courses",
  TeacherGrades = "/teacher/grades",
  Contracts = "/contracts",
  Teachers = "/teachers",
  StudentsReport = "/reports/students",
}